Make a refresher for when a package builds
When on openticket domain and we are working on a frontend package
while it is in watch mode, the dashboard reloads before the
package has fully build. Therefore we reload the page when we have
a page that is not loaded yet (which is a result of the package
still being compiled).